:recycle: refactor(notification): move notification domain from domain module to batch module
