+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/domain/Notification.java
@@ -1,34 +1,29 @@
-package com.depromeet.notification;
+package com.depromeet.domains.notification.domain;
 
-import com.depromeet.notification.vo.OsType;
+import java.time.LocalDateTime;
 import java.util.Date;
 import lombok.Builder;
-import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Getter
 @Builder
-@Document(collection = "user_device")
-public class UserDevice {
+@Document("notification")
+public class Notification {
 
     @Id
     private String id;
-    private Long userId;
-    private String deviceToken;
-    private OsType osType;
-    private String osVersion;
+    private String title;
+    private String content;
+    private String type;
+    private User user;
+    private Target target;
+    private LocalDateTime viewedTime;
     private boolean isDeleted;
     @CreatedDate
     private Date createdAt;
     @LastModifiedDate
     private Date modifiedAt;
-
-    public UserDevice updateDeviceToken(String deviceToken) {
-        this.deviceToken = deviceToken;
-        return this;
-    }
 
 }
