+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/domain/User.java
@@ -1,4 +1,4 @@
-package com.depromeet.notification;
+package com.depromeet.domains.notification.domain;
 
 import lombok.Builder;
