+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/domain/vo/OsType.java
@@ -1,0 +1,5 @@
+package com.depromeet.domains.notification.domain.vo;
+
+public enum OsType {
+    IOS, ANDROID
+}
