+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/repository/NotificationRepository.java
@@ -1,6 +1,6 @@
 package com.depromeet.domains.notification.repository;
 
-import com.depromeet.notification.Notification;
+import com.depromeet.domains.notification.domain.Notification;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface NotificationRepository extends MongoRepository<Notification, String> {
