+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/service/NotificationService.java
@@ -5,11 +5,11 @@ import com.depromeet.domains.notification.dto.request.PushRequestDto;
 import com.depromeet.domains.notification.dto.request.TopicPushRequestDto;
 import com.depromeet.domains.notification.repository.NotificationRepository;
 import com.depromeet.domains.notification.repository.UserDeviceRepository;
-import com.depromeet.notification.Notification;
-import com.depromeet.notification.Target;
-import com.depromeet.notification.User;
-import com.depromeet.notification.UserDevice;
-import com.depromeet.notification.vo.Channel;
+import com.depromeet.domains.notification.domain.Notification;
+import com.depromeet.domains.notification.domain.Target;
+import com.depromeet.domains.notification.domain.User;
+import com.depromeet.domains.notification.domain.UserDevice;
+import com.depromeet.domains.notification.domain.vo.Channel;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
