+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/service/PushService.java
@@ -7,7 +7,7 @@ import com.depromeet.common.error.code.FireBaseErrorCode;
 import com.depromeet.common.error.exceptions.ExternalServerException;
 import com.depromeet.domains.notification.fcm.FcmService;
 import com.depromeet.domains.notification.repository.UserDeviceRepository;
-import com.depromeet.notification.UserDevice;
+import com.depromeet.domains.notification.domain.UserDevice;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
