--- backend/streetdrop-domain/src/main/java/com/depromeet/notification/vo/Channel.java
@@ -1,5 +1,0 @@
-package com.depromeet.notification.vo;
-
-public enum Channel {
-    GENERAL
-}
